A1.8 (Archery Update)
----
System
----
-Added Standard Assets
-Added Loads of medieval weaponry sounds
-Added Slowed Effect
-Added Burn Effect
-Reworked spell system into Projectile system
-Updated Visual Effect Controller functionality
-Added Shop decoration pack
-Added Holy Shift water
-Adjusted how roll and damage animations are triggered
-Added FPS counter
-Moved all damage calculation to a static class
-Removed player  manger need to know who struck it
-Fixed player spawn not being set on first spawn
-Reduced Singleton use in code
-Respawn points can now be protected from destroy
-Made shop portal open with coroutine to add delay to teleport
----
Game
------
-Added Ranger Class
-Added 40+ new items for ranger
-Added 3 new passives for ranger
-Shop has be redecorated and new vendors have been added for all classes plus a potion and accessory merchant
-Adjusted wall height so most cannot be seen over
-Minor adjustments to various items
-Buffed warrior base strength from 26 to 35
-Nerfed warrior base intelligence  from 10 to 5
-Nerfed cleric base strength from 15 to 10
-Shop now has a popup when approached

+++ Assets/Scripts/Items/Status Effects/Efect Scripts/StatusEffectSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public abstract class StatusEffectSO : ScriptableObject
 {
-    public abstract IEnumerator DecayTimer(int time,PlayerManger player, Enemys enemy);
+    public abstract IEnumerator DecayTimer(float time,PlayerManger player, Enemys enemy);
     public abstract IEnumerator EffectApplication();
     public abstract void StatusEffect();
     public abstract void ApplyEffect(object enemy);
